Command Line Arguments
—help and —host

+++ src/main/java/Main.java
@@ -1,9 +1,12 @@
 import com.mongodb.client.model.CreateCollectionOptions;
+import org.apache.commons.cli.ParseException;
 import org.bson.Document;
 import com.mongodb.MongoClient;
 import com.mongodb.MongoClientURI;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
+
+import utils.CommandLineOptions;
 
 public class Main {
 
@@ -11,32 +14,48 @@ public class Main {
 
   public static void main(String[] args) {
 
-    // Server connection
-    String mongoUri = "mongodb://127.0.0.1:27017,127.0.0.1:27018,127.0.0.1:27019";
+    try {
+      CommandLineOptions opts = new CommandLineOptions(args);
 
-    MongoClient client = new MongoClient(new MongoClientURI(mongoUri));
+      // Quit after displaying help message
+      if (opts.isHelpOnly()) return;
 
-    MongoDatabase db = client.getDatabase("IoT");
 
-    // Drop DeviceEvents collection and set to capped with max
-    db.getCollection("DeviceEvents").drop();
 
-    // Capped coll 1GB || 5,000 docs
-    db.createCollection("DeviceEvents", new CreateCollectionOptions()
-            .capped(true)
-            .maxDocuments(5000)
-            .sizeInBytes(1073741824));
+      // Server connection
+      String mongoUri = opts.getMongoUri();
 
-    // Collections
-    MongoCollection<Document> collDeviceEvents = db.getCollection("DeviceEvents");
-    MongoCollection<Document> collDevices = db.getCollection("Devices");
+      MongoClient client = new MongoClient(new MongoClientURI(mongoUri));
 
-    // Get Service and Watch!
-    DeviceService deviceSvc = new DeviceService(collDevices, collDeviceEvents);
+      MongoDatabase db = client.getDatabase("IoT");
 
-    deviceSvc.watchOutOfBounds();
+      // Drop DeviceEvents collection and set to capped with max
+      db.getCollection("DeviceEvents").drop();
 
-    deviceSvc.watchFirstGen();
+      // Capped coll 1GB || 5,000 docs
+      db.createCollection("DeviceEvents", new CreateCollectionOptions()
+              .capped(true)
+              .maxDocuments(5000)
+              .sizeInBytes(1073741824));
+
+      // Collections
+      MongoCollection<Document> collDeviceEvents = db.getCollection("DeviceEvents");
+      MongoCollection<Document> collDevices = db.getCollection("Devices");
+
+      // Get Service and Watch!
+      DeviceService deviceSvc = new DeviceService(collDevices, collDeviceEvents);
+
+      deviceSvc.watchOutOfBounds();
+
+      deviceSvc.watchFirstGen();
+
+
+
+    } catch (ParseException e) {
+
+      System.err.println(e.getMessage());
+      return;
+    }
 
   }
 }
